Update runbook graphs
